requirements.txt
linuxpy>=0.9.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
astropy>=5.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
click>=8.1.0
PyTurboJPEG>=1.7.0
